refactor: move cloud API key loading to core, inline api-engine
- Add settings_key_for_provider_api_key, get_stored_api_key_for_provider,
  require_api_key_for_provider in llm client; streamChat uses require path.
- Expose getApiKeyForProvider on MeAiCore for UI checks and loadModel.
- Remove api-engine.ts; fold cloud engine into unified-engine with
  getCore().getApiKeyForProvider for check/loadModel.
- Fix stream_chat map_err to use error_to_js(&e).

modelId getter now returns the catalog id (_modelId) so getModelInfo works.

Made-with: Cursor
